The driver several times improperly implementing the MySQL Client-Server protocol Generally the ERR_Packet parsing is what missed. If the server sending an ERR_Packet it was misinterpreted or not interpreted at all Misinterpretation was in connection handshake task, when the server sent a ERR_packet with `Too many connections` error, it was interpreted as handshake error and misleading the developer
Silencing an error much more dangerous.
For example the PrepareQuery task can hang the server thread if sends to the server a
wrong SQL query.
The server will respond with an ERR_Packet which was simply ignored.
This method have a closure which must be called with a preprared statement for binding
The error condition must be delivered to this closure to be able to fail the promise

Other places was missing error processing as well

TODO:
   Implement the COM_STMT_FETCH command correctly rather than blaming the MYSQL protocol
   Refactor would be welcome, to implement in every Task (workflow) the correct packet flows
   and error handling
   Implement the missing Tasks:
      - COM_STMT_PING  for connection keep alive
      - COM_STMT_BULK_EXECUTE for Bulk insert
      - COM_STMT_SEND_LONG_DATA for big data Insert/Update
      - COM_STMT_QUERY can use LOCAL INFILE syntax for File uploading. Implement this feature