Add delete column support (#217)
* Add delete column support

Added deleteColumns to MySQLAlterTable

This change allows columns to be deleted (DROP).  I will build on this
with a pull request to fluent-mysql to allow
MySQLDatabase:SchemaSupporting to bind FluentMySQLSchema.deleteColumns
and thus allowing full column ADD/DROP support.

* Fixed test class name

Aligned MySQLMigrationTests name with the file name

Sadly only caught this in Linux build…

* DROP column refinements

Updates as per code review.

* Changed DROP specification to include only the column name

Since MySql 8.0 will not accept the table name in the column name specification of DROP [COLUMN] (where as 5.7 would) I changed the code to only include the identifier and not the table name.